Refactor Automapper to utilize Genetics.
